NF: UndoableChangeDeckMulti simplified
